+++ src/upspring/ViewsGroup.h
@@ -19,7 +19,7 @@ public:
 		if (pwnd->iconic ()) {
 			// store view dimensions
 			viewDims.resize (children ());
-			for (int a=0;a<children();a++) {
+			for (uint a=0;a<(uint)children();a++) {
 				fltk::Widget *w = child (a);
 				Rectangle dim(w->x(),w->y(),w->w(),w->h());
 				viewDims[a] = dim;
@@ -27,7 +27,7 @@ public:
 			wasIconic=true;
 		} else if(wasIconic) {
 			// restore view dimensions
-			for (int a=0;a<children() && a<viewDims.size(); a++) {
+			for (uint a=0;a<(uint)children() && a<viewDims.size(); a++) {
 				fltk::Widget *w = child (a);
 				Rectangle& dim=viewDims[a];
 				w->resize(dim.x(),dim.y(),dim.w(),dim.h());
@@ -35,7 +35,7 @@ public:
 			wasIconic=false;
 		} else if (!suppressRescale) { // this is a normal resize, scale all child-widgets
 			float xs=w()/(float)lastWidth, ys=h()/(float)lastHeight;
-			for (int a=0;a<children();a++) {
+			for (uint a=0;a<(uint)children();a++) {
 				fltk::Widget *ch = child (a);
 				ch->resize(ch->x()*xs+0.5f,ch->y()*ys+0.5f,ch->w()*xs+0.5f,ch->h()*ys+0.5f);
 			}
